move identify_triton_stores to triton.py
